# Improve performance of HTML generator