`getWebTreeMapData` takes too much time with `[combined]` bundle. In my big project, with 10000 files (max directory depth: 11) in the combined bundle, it takes ~45s, while the rest of `exploreResults` only takes 3s in total.

This PR is about to move the combined bundle away from `getWebTreeMapData()`, and re-create it by merging the tree data of other bundles. The merging logic also costs ~2s in my project, but it's much less than 45s.