@nikolay-borzov 2020-04-30
Update snapshots using `test:update-snapshots` script